Add Filtering Support
fuse.js was added as a dependency for filtering learning objects. The structure is in place, and filtering is occuring as expected. Next steps are to produce an observable that the AppComponent can subscribe to in order to update the list shown on screen.

The project name was updated to "cube."