Connect to API
The application now connects into an API to fetch the learning object groups.

The select dialog has been removed, but only through commenting. Once the spoof filtering is implemented in it's own component the comments can be removed.

The way that the service and the app component handle filtered groups has been highly refactored.