Break App into Components
Components have been stubbed out for the faces of the cube that will be implemented, and routing has been configured for each of them. The view and functionality of the AppComponent has been moved to the HomeComponent.

A shared module has been created so that all modules can easily use the curriculum group component.

Contributes to #1 by fully implementing search functionality and stubbing
out the components for the faces of the cube.