Update Home Page
The new nav bar has been added in, along with the removal of the browsable cube. Now the home component contains a search bar that pulls up new learning object components that are clickable to open the detail component.

Cyber Topics, Mapping, Modality, and Academics will become present by the means of a filter.

Contributes to #12