Add Featured LO Section to Home
The home component now has a "Featured Learning Objects" section. The section currently displays all Learning Objects in the system. Once more are generated and metrics given, this section will show only the top rated via said metrics.

The search bar has been disabled in this commit to favor a better search function that is upcoming.

Contributes to story 3 of #12