Clean the AuthModule, make UserCards and LearningObjectListing components accessible to the entire application, add inputs to UserCard and update organization list to match.
AuthModule Changes:
We're validating passwords now! (kinda) The system will alert you if your password isn't strong, but will not prevent you from using it.
New buttons, animations, tooltips, etc for increased usability (and beautiful-ness).
Shows a UserCard item on the final step of registration as a preview.

UserCard Changes:
Added a Link input, if this is false the component won't route to the profile.
Added a gravatarInfo input, if this is true, shows an information button that displays a popup linking to gravatar and explaining what it is.