Disable sortBy date as default value
Sort by date is set as the default value on component creation.
The intent is to show the newest learning objects when a user visits the
browse page. I'll have to discuss implementing this change or would it
be better to set the orderBy query param to nothing when a user
provides a text query?

The orderBy query gets carried towards the next page and after sending
the test query, so the exact match is never at
the top of the search results.